==> Notebook/2_ETL_dataset_principal.ipynb <==
<h1>Extraccion, Transformación y Limpieza de Datos</h1>

<h2>Librerías</h2>

import pandas as pd

<h2>Transformación a CSVs</h2>

internet = pd.ExcelFile('../Datasets Originales/Internet.xlsx')

for hojas_excel in internet.sheet_names:
    df = pd.read_excel(internet, sheet_name=hojas_excel)
    
    archivo_CSV = f'{hojas_excel}.csv'
    
    df.to_csv(archivo_CSV, index=False)
    print(f'Hoja {hojas_excel} guardada como {archivo_CSV}')

<h2>Carga de Datos</h2>

provincia_VMD

provincia_VMD = pd.read_csv('../Hojas CSVs Excel/Velocidad % por prov.csv')
provincia_VMD.info()

provincia_VMD.head(5)

totales_VMD

totales_VMD = pd.read_csv('../Hojas CSVs Excel/Totales VMD.csv')
totales_VMD.info()

totales_VMD.head(5)

accesos_x_tecnologia

accesos_x_tecnologia = pd.read_csv('../Hojas CSVs Excel/Accesos Por Tecnología.csv')
accesos_x_tecnologia.info()

accesos_x_tecnologia.head()

totales_accesos_x_tecnologia

totales_accesos_x_tecnologia = pd.read_csv('../Hojas CSVs Excel/Totales Accesos Por Tecnología.csv')
totales_accesos_x_tecnologia.info()

totales_accesos_x_tecnologia.head()

dial_BAf

dial_BAf = pd.read_csv('../Hojas CSVs Excel/Dial-BAf.csv')
dial_BAf.info()

dial_BAf.head()

totales_dial_BAf

totales_dial_BAf= pd.read_csv('../Hojas CSVs Excel/Totales Dial-BAf.csv')
totales_dial_BAf.info()

totales_dial_BAf.head()

penetracion_poblacion

penetracion_poblacion= pd.read_csv('../Hojas CSVs Excel/Penetración-poblacion.csv')
penetracion_poblacion.info()

penetracion_poblacion.head()

penetracion_hogares

penetracion_hogares= pd.read_csv('../Hojas CSVs Excel/Penetracion-hogares.csv')
penetracion_hogares.info()

penetracion_hogares.head()

penetracion_totales

penetracion_totales= pd.read_csv('../Hojas CSVs Excel/Penetracion-totales.csv')
penetracion_totales.info()

penetracion_totales.head(5)

totales_accesos_x_velocidad

totales_accesos_x_velocidad= pd.read_csv('../Hojas CSVs Excel/Totales Accesos por velocidad.csv')
totales_accesos_x_velocidad.info()

totales_accesos_x_velocidad.head(5)

accesos_x_velocidad

accesos_x_velocidad= pd.read_csv('../Hojas CSVs Excel/Accesos por velocidad.csv')
accesos_x_velocidad.info()

accesos_x_velocidad.head(5)

ingresos

ingresos= pd.read_csv('../Hojas CSVs Excel/Ingresos.csv')
ingresos.info()

ingresos.head(5)

<h2>Eliminación de vacíos</h2>

Eliminamos de los datasets los campos que esten nulos. 

accesos_x_tecnologia = accesos_x_tecnologia.dropna()

<h2>Manejo de valores faltantes</h2>

En El dataset de dial_BA faltan dos datos, pero no los voy a borrar ya que perdería datos. </br>
Los datos que faltan son de Formosa y Misiones del año 2018 para dail up. </br>
Para hayarlos voy a restar al Total el valor de la columna Banda Ancha. 

mask = dial_BAf['Dial up'].isna()
dial_BAf.loc[mask, 'Dial up'] = dial_BAf.loc[mask, 'Total'] - dial_BAf.loc[mask, 'Banda ancha fija']

Para el dataser de accesos_por_velocidad faltan valores en la columna de otros para las provincias de La Rioja, Mendoza, Santa Cruz, Tierra del Fuego y Tucumán y para CABA. </br>
Hay un columna que es TOTAL, para poder ingresar los datos correctamente lo que voy a hacer es restar a esta columa los datos de las otras para que me de el resultado que iría. 

#Para el DF de accesos_por_velocidad hago la cuenta para ver el resultado que iría en la columna "Otros"
columnas_velocidades = [
    'HASTA 512 kbps', 
    '+ 512 Kbps - 1 Mbps', 
    '+ 1 Mbps - 6 Mbps', 
    '+ 6 Mbps - 10 Mbps', 
    '+ 10 Mbps - 20 Mbps', 
    '+ 20 Mbps - 30 Mbps', 
    '+ 30 Mbps'
]

accesos_x_velocidad['suma_velocidades'] = accesos_x_velocidad[columnas_velocidades].sum(axis=1)

accesos_x_velocidad['OTROS'] = accesos_x_velocidad['Total'] - accesos_x_velocidad['suma_velocidades']

accesos_x_velocidad.drop(columns=['suma_velocidades'], inplace=True)

<h2>Duplicados</h2>

#Vemos si hay datos duplicados en los DFs
def valores_duplicados(df): 
    suma_duplicados = df.duplicated().sum()
    if suma_duplicados > 0:
        return f"El Dataframe{df} tiene {suma_duplicados} valores repetidos"
    else:
        return f"El Dataframe no tiene valores duplicados"

print(valores_duplicados(provincia_VMD))
print(valores_duplicados(totales_VMD))
print(valores_duplicados(accesos_x_tecnologia))
print(valores_duplicados(totales_accesos_x_tecnologia))
print(valores_duplicados(dial_BAf))
print(valores_duplicados(totales_dial_BAf))
print(valores_duplicados(penetracion_poblacion))
print(valores_duplicados(penetracion_hogares))
print(valores_duplicados(penetracion_totales))
print(valores_duplicados(totales_accesos_x_velocidad))
print(valores_duplicados(accesos_x_velocidad))
print(valores_duplicados(ingresos))


<h2>Transformación en las columnas</h2>

Em los DF que las columnas estén en dtype float los voy a pasar a int para una mayor comodidad a la hora de visaulizar. </br>
En el DF accesos_x_tecnologia el año y el trimestre tienen un *, voy a quitarlo para poder trabajar en la unión sin inconvenientes. </br>
En la columna de año en ingresos hay un error en un dato, en lugar de ser 2023 pone 2033 </br>
En la culumna de año en accesos por velocidad dice año 2024 y tendría que decir año 2023. </br>

accesos_x_tecnologia['Año'] = accesos_x_tecnologia['Año'].str.replace('*', '')
accesos_x_tecnologia['Trimestre'] = accesos_x_tecnologia['Trimestre'].str.replace('*', '')

accesos_x_tecnologia['Año'] = pd.to_numeric(accesos_x_tecnologia['Año'])
accesos_x_tecnologia['Trimestre'] = pd.to_numeric(accesos_x_tecnologia['Trimestre'])

col_a_mod_acc_x_tecnologia = ['ADSL','Cablemodem', 'Fibra óptica', 'Wireless', 'Otros', 'Total']
accesos_x_tecnologia[col_a_mod_acc_x_tecnologia] = accesos_x_tecnologia[col_a_mod_acc_x_tecnologia].astype(int)


totales_VMD = totales_VMD.drop(['Trimestre.1'], axis=1, errors='ignore')
totales_accesos_x_tecnologia = totales_accesos_x_tecnologia.drop(['Periodo'], axis=1, errors='ignore')
totales_dial_BAf = totales_dial_BAf.drop(['Periodo'], axis=1, errors='ignore')
penetracion_totales = penetracion_totales.drop(['Periodo'], axis=1, errors='ignore')
ingresos = ingresos.drop(['Periodo'], axis=1, errors='ignore')

provincia_VMD = provincia_VMD.round(2)
totales_VMD = totales_VMD.round(2)
accesos_x_tecnologia = accesos_x_tecnologia.round(2)
penetracion_poblacion = penetracion_poblacion.round(2)
penetracion_hogares = penetracion_hogares.round(2)
penetracion_totales = penetracion_totales.round(2)
ingresos = ingresos.round(2)

modificar_columnas = ['Año', 'Trimestre']
totales_VMD[modificar_columnas] = totales_VMD[modificar_columnas].astype(int)
ingresos[modificar_columnas] = ingresos[modificar_columnas].astype(int)

dial_BAf['Dial up'] = dial_BAf['Dial up'].astype(int)
columnas_a_modificar = ['HASTA 512 kbps',
       '+ 512 Kbps - 1 Mbps', '+ 1 Mbps - 6 Mbps', '+ 6 Mbps - 10 Mbps',
       '+ 10 Mbps - 20 Mbps', '+ 20 Mbps - 30 Mbps', '+ 30 Mbps', 'OTROS',
       'Total']
accesos_x_velocidad[columnas_a_modificar] = accesos_x_velocidad[columnas_a_modificar].round(0).astype(int)

modificar_VMD = ['Año', 'Trimestre']
ingresos[modificar_VMD] = ingresos[modificar_VMD].astype(int)

ingresos.loc[ingresos['Año'] == 2033, 'Año'] = 2023
accesos_x_velocidad.loc[accesos_x_velocidad['Año'] == 2024, 'Año'] = 2023
totales_accesos_x_velocidad.loc[totales_accesos_x_velocidad['Año'] == 2024, 'Año'] = 2023

<h2>Descarga Datasets Limpios</h2>

provincia_VMD.to_csv('provincia_VMD.csv', index=False)
totales_VMD.to_csv('totales_VMD.csv', index=False)
accesos_x_tecnologia.to_csv('accesos_x_tecnologia.csv', index=False)
dial_BAf.to_csv('dial_BAf.csv', index=False)
totales_dial_BAf.to_csv('totales_dial_BAf.csv', index=False)
penetracion_poblacion.to_csv('penetracion_poblacion.csv', index=False)
penetracion_hogares.to_csv('penetracion_hogares.csv', index=False)
penetracion_totales.to_csv('penetracion_totales.csv', index=False)
accesos_x_velocidad.to_csv('accesos_x_velocidad.csv', index=False)
totales_accesos_x_velocidad.to_csv('totales_accesos_x_velocidad.csv', index=False)
ingresos.to_csv('ingresos.csv', index=False)
totales_accesos_x_tecnologia.to_csv('totales_accesos_x_tecnologia.csv', index=False)
