==> Notebook/3_comb_df_principal.ipynb <==
<h1>Combinación DFs</h1>

Para poder trabajar mejor voy a convinar diferentes hojas para ir creando diferentes dataframes. 

<h2>Librerías</h2>

import pandas as pd

<h2>Carga Datos</h2>

accesos_x_tecnologia = pd.read_csv('../Dataframes limpios/accesos_x_tecnologia.csv')
accesos_x_velocidad = pd.read_csv('../Dataframes limpios/accesos_x_velocidad.csv')
dial_BAf = pd.read_csv('../Dataframes limpios/dial_BAf.csv')
ingresos = pd.read_csv('../Dataframes limpios/ingresos.csv')
penetracion_hogares = pd.read_csv('../Dataframes limpios/penetracion_hogares.csv')
penetracion_poblacion = pd.read_csv('../Dataframes limpios/penetracion_poblacion.csv')
penetracion_totales = pd.read_csv('../Dataframes limpios/penetracion_totales.csv')
provincia_VMD = pd.read_csv('../Dataframes limpios/provincia_VMD.csv')
totales_accesos_x_velocidad = pd.read_csv('../Dataframes limpios/totales_accesos_x_velocidad.csv')
totales_dial_BAf = pd.read_csv('../Dataframes limpios/totales_dial_BAf.csv')
totales_VMD = pd.read_csv('../Dataframes limpios/totales_VMD.csv')
totales_accesos_x_tecnologia =  pd.read_csv('../Dataframes limpios/totales_accesos_x_tecnologia.csv')


<h2>Unión de DFs</h2>

<b>- inf_provincia:</b> <br>
Es la suma de los DFs de accesos_x_velocidad, velocidad_porcen_x_provincia, accesos_x_tecnologia, dial_BAf, penetracion_hogares y penetracion_poblacion

inf_provincia = pd.merge(accesos_x_velocidad, provincia_VMD, on=['Año', 'Trimestre', 'Provincia'])
inf_provincia.rename(columns={'OTROS': 'Otras Velocidades', 'Total': 'Total Velocidades'}, inplace=True)

inf_provincia = pd.merge(inf_provincia, accesos_x_tecnologia, on=['Año', 'Trimestre', 'Provincia'])
inf_provincia.rename(columns={'OTROS': 'Otros Accesos', 'Total': 'Total Accesos'}, inplace=True)

inf_provincia = pd.merge(inf_provincia, dial_BAf, on=['Año', 'Trimestre', 'Provincia'])
inf_provincia.rename(columns={'Total': 'Total dial_BAf'}, inplace=True)

inf_provincia = pd.merge(inf_provincia, penetracion_hogares, on=['Año', 'Trimestre', 'Provincia'])

inf_provincia = pd.merge(inf_provincia, penetracion_poblacion, on=['Año', 'Trimestre', 'Provincia'])

inf_provincia.head(5)

inf_provincia.columns

<b>- inf_nacional:</b> <br>
Es la suma de los DFs de totales_accesos_x_velocidad, totales_VMD, totales_accesos_x_tecnologia, totales_dial_BAf, penetracion_totales, ingresos

inf_nacional = pd.merge(totales_accesos_x_velocidad, totales_VMD, on=['Año', 'Trimestre'])
inf_nacional.rename(columns={'OTROS': 'Otras Velocidades', 'Total': 'Total Velocidades'}, inplace=True)

inf_nacional = pd.merge(inf_nacional, totales_accesos_x_tecnologia, on=['Año', 'Trimestre'])
inf_nacional.rename(columns={'OTROS': 'Otros Accesos', 'Total': 'Total Accesos'}, inplace=True)

inf_nacional = pd.merge(inf_nacional, totales_dial_BAf, on=['Año', 'Trimestre'])
inf_nacional.rename(columns={'Total': 'Total dial_BAf'}, inplace=True)

inf_nacional = pd.merge(inf_nacional, penetracion_totales, on=['Año', 'Trimestre'])

inf_nacional = pd.merge(inf_nacional, ingresos, on=['Año', 'Trimestre'])

inf_nacional.head(5)

inf_nacional.columns

<h2>Descarga DF Combinados</h2>

inf_provincia.to_csv('inf_provincia.csv', index=False)
inf_nacional.to_csv('inf_nacional.csv', index=False)
