==> Notebook/4_KPIs.ipynb <==
<h1>Creación de KPIs</h1>


<h2>Librerías</h2>

import pandas as pd

<h2>Carga Datos</h2>

inf_nacional = pd.read_csv('../Dataset Combinados/inf_nacional.csv')
inf_provincia = pd.read_csv('../Dataset Combinados/inf_provincia.csv')

<b>KPI = (Nuevoacceso - Accesoactual)/Accesoactual)* 100

Creo la columna para los datos del KPI.

inf_provincia = inf_provincia.sort_values(by=['Año', 'Trimestre'])

def calcular_kpi(df):
    df['NuevoAcceso'] = df['Accesos por cada 100 hogares'].shift(-1)
    
    df['AccesoActual'] = df['Accesos por cada 100 hogares']
    
    df['KPI Acceso'] = ((df['NuevoAcceso'] - df['AccesoActual']) / df['AccesoActual']) * 100
    
    return df

KPI_acceso = calcular_kpi(inf_provincia)

KPI_acceso


inf_provincia = inf_provincia['KPI Acceso'].round(2)

<b> KPI = Ingresos por Tipo de Conexión

# Crear columnas porcentuales para cada tipo de conexión
inf_nacional['ADSL (%)'] = inf_nacional['ADSL'] / inf_nacional['Total Accesos'] * 100
inf_nacional['Cablemodem (%)'] = inf_nacional['Cablemodem'] / inf_nacional['Total Accesos'] * 100
inf_nacional['Fibra óptica (%)'] = inf_nacional['Fibra óptica'] / inf_nacional['Total Accesos'] * 100
inf_nacional['Wireless (%)'] = inf_nacional['Wireless'] / inf_nacional['Total Accesos'] * 100
inf_nacional['Otros (%)'] = inf_nacional['Otros'] / inf_nacional['Total Accesos'] * 100

# Calcular los ingresos por tipo de conexión
inf_nacional['Ingresos ADSL'] = inf_nacional['Ingresos (miles de pesos)'] * inf_nacional['ADSL (%)'] / 100
inf_nacional['Ingresos Cablemodem'] = inf_nacional['Ingresos (miles de pesos)'] * inf_nacional['Cablemodem (%)'] / 100
inf_nacional['Ingresos Fibra óptica'] = inf_nacional['Ingresos (miles de pesos)'] * inf_nacional['Fibra óptica (%)'] / 100
inf_nacional['Ingresos Wireless'] = inf_nacional['Ingresos (miles de pesos)'] * inf_nacional['Wireless (%)'] / 100
inf_nacional['Ingresos Otros'] = inf_nacional['Ingresos (miles de pesos)'] * inf_nacional['Otros (%)'] / 100

# Mostrar el resultado
print(inf_nacional[['Año', 'Ingresos ADSL', 'Ingresos Cablemodem', 'Ingresos Fibra óptica', 'Ingresos Wireless', 'Ingresos Otros']])



inf_nacional.to_csv('KPI_Ingreso_x_conexion.csv', index=False)

<b>KPI = (Nuevoingreso - ingresoactual)/ingresoactual)* 100</b>

inf_nacional = inf_nacional.sort_values(by=['Año'])

def calcular_kpi_inflacion(df):
    df['NuevoIngreso'] = df['Ingresos (miles de pesos)'].shift(-1)
    
    df['IngresoActual'] = df['Ingresos (miles de pesos)']
    
    df['KPI Ingreso'] = ((df['NuevoIngreso'] - df['IngresoActual']) / df['IngresoActual']) * 100
    
    return df

# Asegúrate de que estás llamando a la función correcta
KPI_Ingreso = calcular_kpi_inflacion(inf_nacional)

KPI_Ingreso

inf_nacional = inf_nacional['KPI Ingreso'].round(2)

#KPI_acceso.to_csv('KPI_Acceso.csv', index=False)
KPI_Ingreso.to_csv('KPI_Ingreso.csv', index=False)
